added a color pallet feature
colors file can be used to change the color scheme off the text editor, you need to restart the editor for it to take place. plan on adding the feature as a command.

Please optimize if needed. works for now but might screw up something down the road, did test it though.